add Mathieu Taillefumier's patch for eina logging, with small modifications and fixing
ecore_evas_win32 does not build, though. I think that
if we add log support in evas, all the macro must have
different names, because of all the _private.h headers
that are included in all source files (that's the problem
with win32). I'll fix ecore_evas_win32 build later. Or
someone can do it if he wants :-)


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/ecore@42274 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33