ecore: Sleep at least until the next timeout
The g_main_loop was spinning waiting for timers that were
almost ready.  The spinning was due to waking up from
the g_main_loop too early, and then having nothing to do.

To solve this, round the timeout up to the nearest millisecond,
so that when the main loop wakes, the timer will definitely be
ready.

Signed-off-by: Mike McCormack <mj.mccormack@samsung.com>

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/ecore@61028 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33