Fix busy looping with g_main_loop integration
When glib idlers are present, dispatch can be called after almost
no timeout, triggering our idle enterers and exitters to run again.

This causes our idlers and glib's idlers to ping-pong, and
run in a busy loop.

To solve this, only dispatch when the condition we slept on is satisfied.